send: Fix handling of protocol timeouts
When the protocol takes too long to send a message to the remote, send
will kill that protocol and retry later.  However, it never fully deals
with the signal caused by the child process exiting.  So, when it later
resends the message, it appears that the protocol exits immediately,
causing further problems.

The obvious fix is to make sure we deal with the generated SIGCHLD after
killing the protocol.

Thanks Nick Leverton <nick@leverton.org> for catching this.