lib/fdbuf/fdbuf.h: include <unistd.h>
class fdobuf in lib/fdbuf/fdobuf.h makes use of uid_t and gid_t,
but these are defined in <unistd.h> according to POSIX.  With libc's
that adhere strictly to standards, like musl, this breaks the build.

We add <unistd.h> to fdbuf.h which is included by fdbuf.h.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>