Add ability to read SMTP AUTH password from a configuration file
At present, an authentication password is passed to 'smtp' by way of a
command line option '--pass'. This makes it visible to any user who can
examine the process's status, e.g. with the 'ps' tool.

This change allows a password to be read from a nullmailer configuration
file, thereby removing the password from sight.

A new option '--pass-file' is added, whose argument is expected to be the
name of a single line file containing the password. The file should be
readable only by nullmailer.

The change exposed a linking problem which arises from the order in which
the libraries lib/libnullmailer.a and lib/cli++/libcli++.a are presented to
the linker. An adjustment to protocols/Makefile.am has been made to fix this.