protocols/smtp: Fix treating auth failure as a permanent failure
A typo in the SMTP configuration would cause messages to be rejected due
to the AUTH command being failed. This would result in password typos
causing messages to be bounced.

This change adds a new temporary error code specifically for
authentication failures, which will result in retries (with backoff)
instead of bounces.

Thanks to Fejes József <jozsef.fejes@gmail.com> for identifying this
problem.

This also adds tests for SMTP AUTH which had been missing.